refactor(source): type BranchSelect status enum + move URL validators to utils
BranchSelect's status is now a BranchStatus enum instead of a bare string
union. validateGitUrl + looksResolvable move from NewProjectForm into
utils/sources.ts alongside the other source-URL logic. (review feedback)
